Refactor: Backend update of all school references
