+++ preschoolback/src/main/java/com/center/config/EcryptedConfig.java
@@ -1,0 +1,20 @@
+/**
+ * 
+ */
+package com.center.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+
+import com.ulisesbocchio.jasyptspringboot.annotation.EncryptablePropertySource;
+
+/**
+ * @author francisphiri
+ *
+ */
+@Configuration
+@EncryptablePropertySource(name = "qaEncryptedProperties", value="classpath:qa.ecrypted.properties")
+@Profile("qa")
+public class EcryptedConfig {
+	
+}
