+++ preschoolback/src/main/java/com/center/config/repository/RatingAndReviewRepository.java
@@ -1,0 +1,24 @@
+/**
+ * 
+ */
+package com.center.config.repository;
+
+import java.util.List;
+
+import org.springframework.data.repository.CrudRepository;
+
+import com.center.domain.RatingAndReview;
+import com.center.domain.Center;
+import com.center.domain.User;
+
+/**
+ * @author francisphiri
+ *
+ */
+public interface RatingAndReviewRepository extends CrudRepository<RatingAndReview, Long>{
+	
+	List<RatingAndReview> findByUser(User user);
+	List<RatingAndReview> findByCenter(Center center);
+	List<RatingAndReview> findByMessageContaining(String message);
+	List<RatingAndReview> findAll();
+}
