+++ preschoolback/src/main/java/com/center/config/repository/RoleRepository.java
@@ -1,0 +1,16 @@
+/**
+ * 
+ */
+package com.center.config.repository;
+
+import org.springframework.data.repository.CrudRepository;
+
+import com.center.domain.security.Role;
+
+/**
+ * @author francisphiri
+ *
+ */
+public interface RoleRepository extends CrudRepository<Role, Long>{
+	
+}
