+++ preschoolback/src/main/java/com/center/config/repository/UserRepository.java
@@ -1,0 +1,20 @@
+/**
+ * 
+ */
+package com.center.config.repository;
+
+import java.util.List;
+
+import org.springframework.data.repository.CrudRepository;
+
+import com.center.domain.User;
+
+/**
+ * @author francisphiri
+ *
+ */
+public interface UserRepository extends CrudRepository<User, Long>{
+	User findByUsername(String username);
+	User findByEmail(String email);
+	List<User> findAll();
+}
