+++ preschoolback/src/main/java/com/center/resource/CenterResource.java
@@ -1,0 +1,146 @@
+/**
+ * 
+ */
+package com.center.resource;
+
+import java.io.BufferedOutputStream;
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.multipart.MultipartHttpServletRequest;
+
+import com.center.domain.Center;
+import com.center.service.CenterService;
+
+/**
+ * @author francisphiri
+ *
+ */
+
+@RestController
+@RequestMapping("/center")
+public class CenterResource {
+	
+	@Autowired
+	private CenterService centerService;
+	
+	@RequestMapping (value="/add", method=RequestMethod.POST)
+	public Center addCenterPost(@RequestBody Center center) {
+		return centerService.save(center);
+	}
+	
+	@RequestMapping(value="/add/image", method=RequestMethod.POST)
+	public ResponseEntity upload(
+			@RequestParam("id") Long id,
+			HttpServletResponse response, HttpServletRequest request){
+		try {
+			Center center = centerService.findOne(id);
+			MultipartHttpServletRequest multipartRequest = (MultipartHttpServletRequest) request;
+			Iterator<String> it = multipartRequest.getFileNames();
+			MultipartFile multipartFile = multipartRequest.getFile(it.next());
+			String fileName = id+".png";
+			
+			
+			byte[] bytes = multipartFile.getBytes();
+			BufferedOutputStream stream = new BufferedOutputStream(new FileOutputStream(new File("src/main/resources/static/image/center/"+fileName)));
+			stream.write(bytes);
+			stream.close();
+			
+			return new ResponseEntity("Upload Success!", HttpStatus.OK);
+		} catch (Exception e) {
+			e.printStackTrace();
+			return new ResponseEntity("Upload failed!", HttpStatus.BAD_REQUEST);
+		}
+	}
+	
+	@RequestMapping(value="/update/image", method=RequestMethod.POST)
+	public ResponseEntity updateImagePost(
+			@RequestParam("id") Long id,
+			HttpServletResponse response, HttpServletRequest request
+			){
+		try {
+			Center center = centerService.findOne(id);
+			MultipartHttpServletRequest multipartRequest = (MultipartHttpServletRequest) request;
+			Iterator<String> it = multipartRequest.getFileNames();
+			MultipartFile multipartFile = multipartRequest.getFile(it.next());
+			String fileName = id+".png";
+			
+			Files.delete(Paths.get("src/main/resources/static/image/center/"+fileName));
+			
+			byte[] bytes = multipartFile.getBytes();
+			BufferedOutputStream stream = new BufferedOutputStream(
+					new FileOutputStream(new File("src/main/resources/static/image/center/"+fileName)));
+			stream.write(bytes);
+			stream.close();
+			
+			return new ResponseEntity("Upload Success!", HttpStatus.OK);
+		} catch (Exception e) {
+			e.printStackTrace();
+			return new ResponseEntity("Upload failed!", HttpStatus.BAD_REQUEST);
+		}
+	}
+	
+	@RequestMapping("/centerList")
+	public List<Center> getCenterList() {
+		List<Center> allCenters = centerService.findAll();
+		System.out.println(allCenters);
+		return allCenters; //.subList(0, 5);
+	}
+	
+	@RequestMapping("/wishList/{wishList}")
+	public List<Center> getWishList(@PathVariable("wishList") String wishList) {
+		System.out.println(" here : "+wishList);
+		return centerService.findWishList(wishList); 
+	}
+	
+	@RequestMapping(value="/update", method=RequestMethod.POST)
+	public Center updateCenterPost(@RequestBody Center center) {
+		return centerService.save(center);
+	}
+	
+	@RequestMapping(value="/remove", method=RequestMethod.POST)
+	public ResponseEntity remove(
+			@RequestBody String id
+			) throws IOException {
+		centerService.removeOne(Long.parseLong(id));
+		String fileName = id+".png";
+		
+		Files.delete(Paths.get("src/main/resources/static/image/center/"+fileName));
+		
+		return new ResponseEntity("Remove Success!", HttpStatus.OK);
+	}
+	
+	@RequestMapping("/{id}")
+	public Center getBook(@PathVariable("id") Long id){
+		Center book = centerService.findOne(id);
+		return book;
+	}
+	
+	@RequestMapping(value="/searchCenter", method=RequestMethod.POST)
+	public List<Center> searchBook (@RequestBody String keyword) {
+		List<Center> bookList = centerService.blurrySearch(keyword);
+		
+		return bookList;
+	}
+	
+	
+}
