+++ preschoolback/src/main/java/com/center/service/UserSecurityService.java
@@ -1,0 +1,39 @@
+/**
+ * 
+ */
+package com.center.service;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+import com.center.config.repository.UserRepository;
+import com.center.domain.User;
+
+/**
+ * @author francisphiri
+ *
+ */
+@Service
+public class UserSecurityService implements UserDetailsService{
+	
+	private static final Logger LOG = LoggerFactory.getLogger(UserSecurityService.class);
+	
+	@Autowired 
+	private UserRepository userRepository;
+	
+	@Override
+	public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+		User user = userRepository.findByUsername(username);
+		if(null == user) {
+			LOG.warn("Username {} not found", username);
+			throw new UsernameNotFoundException("Username "+username+" not found");
+		}
+		return user;
+	}
+	
+}
