--- preschoolback/src/main/java/com/preschool/config/HttpSessionConfig.java
@@ -1,22 +1,0 @@
-/**
- * 
- */
-package com.preschool.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-import org.springframework.session.data.redis.config.annotation.web.http.EnableRedisHttpSession;
-
-/**
- * @author francisphiri
- *
- */
-@EnableRedisHttpSession
-public class HttpSessionConfig {
-	@Bean
-	public LettuceConnectionFactory connectionFactory() {
-		//return new LettuceConnectionFactory("http://eyas-redis.ajkq9r.ng.0001.usw1.cache.amazonaws.com",6379);
-		return new LettuceConnectionFactory();//currently using embedded redis
-	}
-}
-
