--- preschoolback/src/main/java/com/preschool/config/RequestFilter.java
@@ -1,64 +1,0 @@
-/**
- * 
- */
-package com.preschool.config;
-
-import java.io.IOException;
-
-import javax.servlet.Filter;
-import javax.servlet.FilterChain;
-import javax.servlet.FilterConfig;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import org.springframework.core.Ordered;
-import org.springframework.core.annotation.Order;
-import org.springframework.stereotype.Component;
-
-/**
- * @author francis
- *
- */
-@Component
-@Order(Ordered.HIGHEST_PRECEDENCE)
-public class RequestFilter implements Filter{
-
-	public void doFilter(ServletRequest req, ServletResponse res, FilterChain chain) {
-        HttpServletResponse response = (HttpServletResponse) res;
-        HttpServletRequest request = (HttpServletRequest) req;
-       
-        response.setHeader("Access-Control-Allow-Origin", "http://localhost:4200");
-        //response.setHeader("Access-Control-Allow-Origin", "http://eyaseyes.com");
-        
-        response.setHeader("Access-Control-Allow-Methods", "POST, PUT, GET, OPTIONS, DELETE");
-        response.setHeader("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
-        response.setHeader("Access-Control-Max-Age", "3600");
-        response.setHeader("Access-Control-Allow-Credentials", "true");
-
-        if (!(request.getMethod().equalsIgnoreCase("OPTIONS"))) {
-            try {
-                chain.doFilter(req, res);
-            } catch(Exception e) {
-                e.printStackTrace();
-            }
-        } else {
-            System.out.println("Pre-flight");
-            response.setHeader("Access-Control-Allow-Methods", "POST,GET,DELETE,OPTIONS");
-            response.setHeader("Access-Control-Max-Age", "3600");
-            response.setHeader("Access-Control-Allow-Headers", "authorization, content-type," +
-                    "access-control-request-headers,access-control-request-method,accept,origin,authorization,x-requested-with,"
-                    + "x-auth-token");
-            response.setStatus(HttpServletResponse.SC_OK);
-        }
-
-    }
-
-    public void init(FilterConfig filterConfig) {}
-
-    public void destroy() {}
-
-
-}
