--- preschoolback/src/main/java/com/preschool/config/repository/RatingAndReviewRepository.java
@@ -1,24 +1,0 @@
-/**
- * 
- */
-package com.preschool.config.repository;
-
-import java.util.List;
-
-import org.springframework.data.repository.CrudRepository;
-
-import com.preschool.domain.RatingAndReview;
-import com.preschool.domain.School;
-import com.preschool.domain.User;
-
-/**
- * @author francisphiri
- *
- */
-public interface RatingAndReviewRepository extends CrudRepository<RatingAndReview, Long>{
-	
-	List<RatingAndReview> findByUser(User user);
-	List<RatingAndReview> findBySchool(School school);
-	List<RatingAndReview> findByMessageContaining(String message);
-	List<RatingAndReview> findAll();
-}
