--- preschoolback/src/main/java/com/preschool/config/repository/UserRepository.java
@@ -1,20 +1,0 @@
-/**
- * 
- */
-package com.preschool.config.repository;
-
-import java.util.List;
-
-import org.springframework.data.repository.CrudRepository;
-
-import com.preschool.domain.User;
-
-/**
- * @author francisphiri
- *
- */
-public interface UserRepository extends CrudRepository<User, Long>{
-	User findByUsername(String username);
-	User findByEmail(String email);
-	List<User> findAll();
-}
