--- preschoolback/src/main/java/com/preschool/service/SchoolService.java
@@ -1,30 +1,0 @@
-/**
- * 
- */
-package com.preschool.service;
-
-import java.util.List;
-
-import com.preschool.domain.RatingAndReview;
-import com.preschool.domain.School;
-
-/**
- * @author francisphiri
- *
- */
-public interface SchoolService {
-	
-	List<School> findAll();
-	
-	List<School> findWishList(String iDs);
-	
-	School findOne(Long id);
-	
-	School save(School school);
-	
-	List<School> blurrySearch(String title);
-	
-	void removeOne(Long id);
-
-	void updateAverageSchoolRating(School school,RatingAndReview ratingAndReview);
-}
