--- preschoolback/src/main/java/com/preschool/utility/MailConstructor.java
@@ -1,65 +1,0 @@
-/**
- * 
- */
-package com.preschool.utility;
-
-import java.util.Locale;
-
-import javax.mail.internet.InternetAddress;
-import javax.mail.internet.MimeMessage;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.env.Environment;
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.MimeMessageHelper;
-import org.springframework.mail.javamail.MimeMessagePreparator;
-import org.springframework.stereotype.Component;
-import org.thymeleaf.TemplateEngine;
-import org.thymeleaf.context.Context;
-
-import com.preschool.domain.User;
-
-/**
- * @author francisphiri
- *
- */
-@Component
-public class MailConstructor {
-	
-	@Autowired
-	private Environment env;
-	
-	@Autowired
-	private TemplateEngine templateEngine;
-	
-	public SimpleMailMessage constructNewUserEmail(User user, String password) {
-		String message="\nPlease use the following credentials to log in and edit your personal information including your own password."
-				+ "\nUsername:"+user.getUsername()+"\nPassword:"+password;
-		
-		SimpleMailMessage email = new SimpleMailMessage();
-		email.setTo(user.getEmail());
-		email.setSubject("Eyas Eyes - New User");
-		email.setText(message);
-		email.setFrom(env.getProperty("support.email"));
-		return email;
-	}
-	
-	public MimeMessagePreparator constructOrderConfirmationEmail (User user, Locale locale) {
-		Context context = new Context();
-		context.setVariable("user", user);
-		String text = templateEngine.process("orderConfirmationEmailTemplate", context);
-		
-		MimeMessagePreparator messagePreparator = new MimeMessagePreparator() {
-			@Override
-			public void prepare(MimeMessage mimeMessage) throws Exception {
-				MimeMessageHelper email = new MimeMessageHelper(mimeMessage);
-				email.setTo(user.getEmail());
-				email.setSubject("School Registration Confirmation");
-				email.setText(text,true);
-				email.setFrom(new InternetAddress("nursery.express@gmail.com"));
-			}
-		};
-		
-		return messagePreparator;
-	}
-}
