+++ preschoolback/src/test/java/com/center/PreschoolApplicationTests.java
@@ -1,0 +1,16 @@
+package com.center;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest(classes = EyasApplication.class)
+public class PreschoolApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
