--- preschoolback/src/test/java/com/preschool/PreschoolApplicationTests.java
@@ -1,16 +1,0 @@
-package com.preschool;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest(classes = PreschoolApplication.class)
-public class PreschoolApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
